Move mkb to separate crate for binding generation
This allows build.rs to run more selectively for binding generation than
the main build.rs since binding generation is slow.
