+++ crates/mkb/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "mkb"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+num_enum = { version = "0.7.3", default-features = false }
+
+[build-dependencies]
+bindgen = "0.70.1"
