+++ crates/mkb/src/lib.rs
@@ -1,0 +1,8 @@
+#![no_std]
+
+#[allow(dead_code)]
+#[allow(non_upper_case_globals)]
+#[allow(non_camel_case_types)]
+#[allow(non_snake_case)]
+pub mod mkb;
+pub mod mkb_suppl;
