+++ crates/mkb/src/mkb2_ghidra.h
