+++ crates/smb2_practice_mod/src/mods/camera.rs
@@ -1,7 +1,7 @@
+use mkb::mkb;
 use num_enum::TryFromPrimitive;
 
 use crate::app::AppContext;
-use crate::mkb;
 use crate::patch;
 use crate::systems::pref::Pref;
 use crate::systems::pref::{self, U8Pref};
