+++ crates/smb2_practice_mod/src/mods/cmseg.rs
@@ -1,10 +1,11 @@
 use core::{ffi::c_long, ptr::null_mut};
 
+use mkb::mkb;
 use num_enum::TryFromPrimitive;
 
 use crate::{
     app::AppContext,
-    hook, mkb,
+    hook,
     systems::{
         draw,
         pref::{BoolPref, Pref, U8Pref},
