+++ crates/smb2_practice_mod/src/mods/gotostory.rs
@@ -1,4 +1,5 @@
-use crate::{app::AppContext, mkb};
+use crate::app::AppContext;
+use mkb::mkb;
 
 #[derive(Default)]
 enum State {
