+++ crates/smb2_practice_mod/src/mods/ilbattle.rs
@@ -1,9 +1,10 @@
 use arrayvec::ArrayString;
+use mkb::mkb;
 use num_enum::TryFromPrimitive;
 
 use crate::{
     app::AppContext,
-    cstr_buf, fmt, mkb,
+    cstr_buf, fmt,
     systems::{
         binds::Binds,
         draw,
