+++ crates/smb2_practice_mod/src/mods/ilmark.rs
@@ -1,10 +1,11 @@
 use core::ffi::c_char;
 
 use arrayvec::ArrayString;
+use mkb::mkb;
 
 use crate::{
     app::AppContext,
-    cstr_buf, mkb,
+    cstr_buf,
     systems::{
         draw,
         pref::{BoolPref, Pref},
