+++ crates/smb2_practice_mod/src/mods/sfx.rs
@@ -1,6 +1,8 @@
+use mkb::mkb;
+
 use core::ffi::c_long;
 
-use crate::{app::AppContext, hook, mkb, systems::pref::BoolPref};
+use crate::{app::AppContext, hook, systems::pref::BoolPref};
 
 hook!(SoftStreamStartHook, looping_state: u32, g_bgm_id: mkb::BgmTrack, param_3: u32 => c_long,
     mkb::SoftStreamStart, |_looping_state, _g_bgm_id, _param_3, _cx| {
