+++ crates/smb2_practice_mod/src/mods/timer.rs
@@ -1,6 +1,7 @@
+use mkb::mkb;
+
 use crate::{
     app::AppContext,
-    mkb,
     systems::{draw, pref::BoolPref},
     utils::{memstore::MemStore, timerdisp},
 };
