+++ crates/smb2_practice_mod/src/mods/unlock.rs
@@ -1,5 +1,6 @@
+use mkb::mkb;
+
 use crate::app::AppContext;
-use crate::mkb;
 use crate::systems::pref::{BoolPref, Pref};
 use crate::utils::misc::for_c_arr;
