+++ crates/smb2_practice_mod/src/systems/binds.rs
@@ -1,4 +1,6 @@
-use crate::{app::AppContext, mkb};
+use mkb::mkb;
+
+use crate::app::AppContext;
 use arrayvec::ArrayString;
 
 use super::pad::{Pad, Prio};
