+++ crates/smb2_practice_mod/src/systems/pref.rs
@@ -1,6 +1,7 @@
 extern crate alloc;
+use ::mkb::mkb_suppl::CARDResult;
 
-use crate::{app::AppContext, cstr, fmt, mkb_suppl::CARDResult, systems::draw::NotifyDuration};
+use crate::{app::AppContext, cstr, fmt, systems::draw::NotifyDuration};
 use alloc::vec;
 use alloc::vec::Vec;
 use num_enum::TryFromPrimitive;
