+++ crates/smb2_practice_mod/src/utils/log.rs
@@ -5,12 +5,12 @@
 macro_rules! log {
     ($fmt:literal $(, $arg:expr)* $(,)?) => {{
         unsafe {
-            $crate::mkb::OSReport(c"[pracmod] ".as_ptr() as *mut _);
+            mkb::mkb::OSReport(c"[pracmod] ".as_ptr() as *mut _);
             // I'm not sure it's possible to expand arg in a safe context
             #[allow(clippy::macro_metavars_in_unsafe)]
             let fmt: &core::ffi::CStr = $fmt;
-            $crate::mkb::OSReport(fmt.as_ptr() as *mut _ $(, $arg)*);
-            $crate::mkb::OSReport(c"\n".as_ptr() as *mut _);
+            mkb::mkb::OSReport(fmt.as_ptr() as *mut _ $(, $arg)*);
+            mkb::mkb::OSReport(c"\n".as_ptr() as *mut _);
         }
     }};
 }
