+++ crates/smb2_practice_mod/src/utils/string_utils.rs
@@ -4,7 +4,7 @@ macro_rules! fmt_buf {
         unsafe {
             let fmt: &core::ffi::CStr = $fmt;
             let ptr = $buf.as_mut_ptr();
-            let written_bytes = $crate::mkb::sprintf(ptr as *mut _, fmt.as_ptr() as *mut _ $(, $arg)*);
+            let written_bytes = ::mkb::mkb::sprintf(ptr as *mut _, fmt.as_ptr() as *mut _ $(, $arg)*);
             assert!(written_bytes >= 0);
             // Hopefully this panics before whatever we overwrote causes problems
             assert!((written_bytes as usize) < $buf.capacity());
