+++ crates/smb2_practice_mod/src/utils/timerdisp.rs
@@ -1,4 +1,5 @@
-use crate::{cstr, fmt, mkb, systems::draw};
+use crate::{cstr, fmt, systems::draw};
+use mkb::mkb;
 
 const SECOND_FRAMES: u32 = 60;
 const MINUTE_FRAMES: u32 = SECOND_FRAMES * 60;
